admin_panel

+++ bot/handlers/admin/admin_panel.py
@@ -4,14 +4,15 @@ from datetime import datetime
 from telebot.types import Message
 from bot import bot, logger
 
-from AI import settings
 from bot.texts import ADMIN_PANEL_TEXT
-from bot.models import User
+from bot.models import User, Transaction
 
-def admin_permission(func):
+
+'''def admin_permission(func):
     """
     Checking user for admin permission to access the function.
     """
+
     @wraps(func)
     def wrapped(message: Message) -> None:
         user_id = message.from_user.id
@@ -21,12 +22,27 @@ def admin_permission(func):
             logger.warning(f'Попытка доступа к админ панели от {user_id}')
             return
         return func(message)
+
     return wrapped
 
-@admin_permission
+
+@admin_permission'''
 def admin_panel(message: Message):
+    print("aa")
+    return
+    if not Transaction.objects.filter(is_addition=False).exists():
+        bot.send_message(chat_id=message.chat.id, text="К сожалению статистика отсутствует")
+        return
+    print("ok")
+    transactions = Transaction.objects.filter(is_addition=False)
+    total_sum = float()
+    tokens_given = float()
+    for transaction in transactions:
+        total_sum += transaction.cash
+        tokens_given += transaction.token_given
     user_id = message.from_user.id
     bot.delete_message(chat_id=message.chat.id, message_id=message.id)
     bot.send_message(chat_id=message.chat.id,
-                     text=f"{ADMIN_PANEL_TEXT}, {user_id}\n\nВот статистика за {datetime.now().month}:\n\nПотрачено токенов пользователями: {settings.token_counter}",
-    )
+                     text=f"{ADMIN_PANEL_TEXT}, {user_id}\n\nВот статистика за {datetime.now().month}:\n\nПотрачено "
+                          f"токенов пользователями: {tokens_given}\n\nПолучено прибыли: {total_sum}",
+                     )
